test_objects/test_cart_page.py
import pytest
from selenium import webdriver
from pages.cart_page import CartPage

# Set up a fixture to initialize the web driver and open the web page
@pytest.fixture
def browser():
    driver = webdriver.Chrome()  # You can use any other web driver you prefer
    driver.implicitly_wait(10)
    yield driver
    driver.quit()

def test_get_cart_items_count(browser):
    cart_page = CartPage(browser)
    cart_page.open("https://store.ferrari.com/us/OnePageCheckout/Cart")  # Replace with your actual cart page URL

    cart_items_count = cart_page.get_cart_items_count()

    # Add assertions to verify the count of items in the cart, e.g., assert cart_items_count == expected_count.

def test_proceed_to_checkout(browser):
    cart_page = CartPage(browser)
    cart_page.open("https://example.com/cart")

    cart_page.proceed_to_checkout()

    # Add assertions to verify that clicking on the "Checkout" button navigates to the checkout page or takes the expected action.

def test_get_empty_cart_message(browser):
    cart_page = CartPage(browser)
    cart_page.open("https://example.com/cart")

    empty_cart_message = cart_page.get_empty_cart_message()

    # Add assertions to verify that the empty cart message is displayed correctly, e.g., assert "Your cart is empty" in empty_cart_message.

# Add more test cases as needed
